Move dropzone to ema component
