+++ core-web/libs/portlets/dot-ema/src/lib/edit-ema-editor/components/ema-page-dropzone/ema-page-dropzone.component.ts
@@ -8,7 +8,7 @@ import {
     Output
 } from '@angular/core';
 
-import { ActionPayload } from '../../shared/models';
+import { ActionPayload } from '../../../shared/models';
 
 interface Contentlets {
     x: number;
